Fix use-after-free
Summary:
We now detect if IncidenceAttachment is deleted while the
AttachmentEditDialog is shown. There were three potential crashes
(deletion of item; access to mAttachmentView; call checkDirtyStatus);
the first of them actually was met in the wild in Bug 356053. Fixing
only the deletion of item led to the other two.

BUG: 356053
FIXED-IN: 5.6.1

Test Plan:
Compiled; repeated the steps described in Bug 356053. The crash
happens without the patch; doesn't happen with the patch.

Reviewers: #kde_pim, dvratil

Reviewed By: #kde_pim, dvratil

Subscribers: winterz

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D7591