Fix keyboard editting of start/end times of VTODOs
Summary:
Editting start/end times of VTODOs using only the keyboard did not
trigger a check whether the VTODO is dirty. If other fields were updated
along with start/end times, everything was fine, but only editting those
times was simply ignored. The patch causes a dirty check for keyboard
edits of the corresponding KTimeComboBoxes.

BUG: 391769
FIXED-IN: 5.9.3

Test Plan:
I changed start and end times of a test VTODO. Without the patch,
the change was not saved when hitting OK, nor did I hit my dirty-check
breakpoint. With the patch, everything works as expected: Keyboard editting
results in an updated VTODO, as verified via the Month View tooltip and by
opening the IncidenceEditor again for that VTODO. VEVENTs do not have this
problem even without the patch.

Reviewers: #kde_pim, winterz

Reviewed By: #kde_pim, winterz

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D16114