IncidenceAttachment: don't allow running executables
This only sounds like a recipe for security issues, with no good use
case. Looking at the history in the kde4 kdepim repo is difficult
because of rewrites of the class, but it really looks like this was
a boolean-trap victim, the initial "true" in runURL() was about
'tempFile', not about 'runExecutables'.