Do not create duplicate tags
If an event with categories is cut & pasted, the pasted event has the
same categories as the original but does not have the same tag-item
relationships in Akonadi.  (This is not the only way that categories
and tags go out of sync.)

When the incidence editor opens the pasted event, it creates new GENERIC
tags for the categories, duplicating the existing tag.  Then it inserts
the tag in the editor, thus making it dirty and causing unnecessary
"Do you really want to cancel?" dialogs.

This patch creates PLAIN tags, merges them with existing tags, and
does not dirty the editor.

BUG: 441846