Synchronize "floating" status of date-times
RFC5545 states that the DTEND and DUE properties must be specified with
local ("floating") times if and only if DTSTART is local.  This patch
ensures that the incidence editor obeys that rule by setting the two
timezone combo boxes to the same time zone if either of them changes to
or from the "floating" time zone.

- `updateStartSpec()` and `updateEndSpec()` synchronize the combo boxes.
- `enableStartEdit()` and `enableEndEdit()` no longer change time zones
  when a todo's start or due time is enabled or disabled.  Disabling
  one would set both to "floating"!
- `identical()` ensures that all changes to time zones are detected.

@dfaure