Fix crash when creating all-day to-dos
Creating all-day to-dos by right-clicking in the month view or in the
all-day area of the agenda view caused an assertion failure, "editor
shouldn't be dirty", because the timeSpec of an all-day event is not
rigorously defined.  This patch codes around the problem and imitates
the "dirty" logic for events.

@dfaure