Add UI to sign/encrypt iTIP invitations
Previously this was governed purely by the information in organizer's
Identity (plus attendees crypto preferences in KAddressBook, if any),
now it's possible for the user to explicitly choose what to do.

This is the first step, long-term I would like to have the same UX
for the attendees as we have for recipients in KMail composer.