Reset AttendeeTableModel when settings attendees
Reset the model instead of emitting layoutChanged, because setting
attendees may actually result in different number of rows in the model
than before, which should not happen during layout change. This seems
to have mostly worked, but caused crash inside QSortFilterProxyModel
when the number of rows have actually changed (most likely connected
to the trailing empty row).

BUG: 428888
FIXED-IN: 23.08.3